+++ src/cli/index.js
@@ -1,7 +1,7 @@
 const glob = require("glob");
 const _get = require("lodash/get");
 const main = require("../main");
-const bus = require("../main/bus");
+const bus = require("../utl/bus");
 const parseTSConfig = require("./parse-ts-config");
 const parseBabelConfig = require("./parse-babel-config");
 const getResolveConfig = require("./get-resolve-config");
