+++ src/main/index.js
@@ -4,8 +4,8 @@ const enrich = require("../enrich");
 const cruiseResultSchema = require("../schema/cruise-result.schema.json");
 const meta = require("../extract/transpile/meta");
 const report = require("../report");
+const bus = require("../utl/bus");
 const normalizeFilesAndDirectories = require("./files-and-dirs/normalize");
-const bus = require("./bus");
 const validateRuleSet = require("./rule-set/validate");
 const normalizeRuleSet = require("./rule-set/normalize");
 const validateOptions = require("./options/validate");
